+++ common/cache/domainCache_test.go
@@ -25,7 +25,9 @@ import (
 	"testing"
 	"time"
 
+	gomock "github.com/golang/mock/gomock"
 	"github.com/pborman/uuid"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
@@ -270,6 +272,61 @@ func (s *domainCacheSuite) TestGetDomain_NonLoaded_GetByID() {
 	entryByID, err = s.domainCache.GetDomainByID(domainRecord.Info.ID)
 	s.Nil(err)
 	s.Equal(entry, entryByID)
+}
+
+func Test_IsActiveIn(t *testing.T) {
+	tests := []struct {
+		msg              string
+		isGlobalDomain   bool
+		currentCluster   string
+		activeCluster    string
+		failoverDeadline *int64
+		expectIsActive   bool
+		expectedErr      error
+	}{
+		{
+			msg:            "local domain",
+			isGlobalDomain: false,
+			expectIsActive: true,
+		},
+		{
+			msg:              "global pending active domain",
+			isGlobalDomain:   true,
+			failoverDeadline: common.Int64Ptr(time.Now().Unix()),
+			expectedErr:      &types.DomainNotActiveError{Message: "Domain: test-domain is pending active in cluster: .", DomainName: "test-domain", CurrentCluster: "", ActiveCluster: ""},
+		},
+		{
+			msg:            "global domain on active cluster",
+			isGlobalDomain: true,
+			currentCluster: "A",
+			activeCluster:  "A",
+			expectIsActive: true,
+		},
+		{
+			msg:            "global domain on passive cluster",
+			isGlobalDomain: true,
+			currentCluster: "A",
+			activeCluster:  "B",
+			expectedErr:    &types.DomainNotActiveError{Message: "Domain: test-domain is active in cluster: B, while current cluster A is a standby cluster.", DomainName: "test-domain", CurrentCluster: "A", ActiveCluster: "B"},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.msg, func(t *testing.T) {
+			domain := NewDomainCacheEntryForTest(
+				&persistence.DomainInfo{Name: "test-domain"},
+				nil,
+				tt.isGlobalDomain,
+				&persistence.DomainReplicationConfig{ActiveClusterName: tt.activeCluster},
+				0,
+				tt.failoverDeadline,
+			)
+
+			isActive, err := domain.IsActiveIn(tt.currentCluster)
+
+			assert.Equal(t, tt.expectIsActive, isActive)
+			assert.Equal(t, tt.expectedErr, err)
+		})
+	}
 }
 
 func (s *domainCacheSuite) TestRegisterCallback_CatchUp() {
@@ -535,7 +592,7 @@ func (s *domainCacheSuite) TestGetTriggerListAndUpdateCache_ConcurrentAccess() {
 }
 
 func (s *domainCacheSuite) buildEntryFromRecord(record *persistence.GetDomainResponse) *DomainCacheEntry {
-	newEntry := newDomainCacheEntry()
+	newEntry := &DomainCacheEntry{}
 	newEntry.info = &*record.Info
 	newEntry.config = &*record.Config
 	newEntry.replicationConfig = &persistence.DomainReplicationConfig{
@@ -635,4 +692,57 @@ func Test_IsSampledForLongerRetention(t *testing.T) {
 
 	d.info.Data[SampleRateKey] = "invalid-value"
 	require.False(t, d.IsSampledForLongerRetention(wid))
+}
+
+func Test_GetActiveDomainByID(t *testing.T) {
+	nonExistingUUID := uuid.New()
+	activeDomainUUID := uuid.New()
+	passiveDomainUUID := uuid.New()
+
+	activeDomain := NewGlobalDomainCacheEntryForTest(&persistence.DomainInfo{ID: activeDomainUUID, Name: "active"}, nil, &persistence.DomainReplicationConfig{ActiveClusterName: "A"}, 0)
+	passiveDomain := NewGlobalDomainCacheEntryForTest(&persistence.DomainInfo{ID: passiveDomainUUID, Name: "passive"}, nil, &persistence.DomainReplicationConfig{ActiveClusterName: "B"}, 0)
+
+	tests := []struct {
+		msg          string
+		domainID     string
+		expectDomain *DomainCacheEntry
+		expectedErr  error
+	}{
+		{
+			msg:         "invalid UUID",
+			domainID:    "invalid",
+			expectedErr: &types.BadRequestError{Message: "Invalid domain UUID."},
+		},
+		{
+			msg:         "non existing domain",
+			domainID:    nonExistingUUID,
+			expectedErr: assert.AnError,
+		},
+		{
+			msg:          "active domain",
+			domainID:     activeDomainUUID,
+			expectDomain: activeDomain,
+		},
+		{
+			msg:          "passive domain",
+			domainID:     passiveDomainUUID,
+			expectDomain: passiveDomain,
+			expectedErr:  &types.DomainNotActiveError{Message: "Domain: passive is active in cluster: B, while current cluster A is a standby cluster.", DomainName: "passive", CurrentCluster: "A", ActiveCluster: "B"},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.msg, func(t *testing.T) {
+			ctrl := gomock.NewController(t)
+			cache := NewMockDomainCache(ctrl)
+			cache.EXPECT().GetDomainByID(nonExistingUUID).Return(nil, assert.AnError).AnyTimes()
+			cache.EXPECT().GetDomainByID(activeDomainUUID).Return(activeDomain, nil).AnyTimes()
+			cache.EXPECT().GetDomainByID(passiveDomainUUID).Return(passiveDomain, nil).AnyTimes()
+
+			domain, err := GetActiveDomainByID(cache, "A", tt.domainID)
+
+			assert.Equal(t, tt.expectDomain, domain)
+			assert.Equal(t, tt.expectedErr, err)
+		})
+	}
 }
