+++ service/history/task/cross_cluster_source_task_executor.go
@@ -27,7 +27,6 @@ import (
 	"fmt"
 
 	"github.com/uber/cadence/common"
-	"github.com/uber/cadence/common/domain"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/metrics"
 	"github.com/uber/cadence/common/persistence"
@@ -553,7 +552,7 @@ func (t *crossClusterSourceTaskExecutor) verifyDomainActive(
 		return ErrTaskPendingActive
 	}
 
-	if isActive, _ := domain.IsActive(entry, t.shard.GetClusterMetadata()); !isActive {
+	if isActive, _ := entry.IsActiveIn(t.shard.GetClusterMetadata().GetCurrentClusterName()); !isActive {
 		// set processing state to invalidated so that a new task can be created
 		t.setTaskState(task, ctask.TaskStatePending, processingStateInvalidated)
 		return nil
