+++ service/history/task/cross_cluster_target_task_executor.go
@@ -26,7 +26,6 @@ import (
 
 	"github.com/uber/cadence/client/history"
 	"github.com/uber/cadence/common"
-	"github.com/uber/cadence/common/domain"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/metrics"
@@ -448,7 +447,7 @@ func (t *crossClusterTargetTaskExecutor) verifyDomainActive(
 		return "", ErrTaskPendingActive
 	}
 
-	if isActive, _ := domain.IsActive(entry, t.shard.GetClusterMetadata()); !isActive {
+	if isActive, _ := entry.IsActiveIn(t.shard.GetClusterMetadata().GetCurrentClusterName()); !isActive {
 		return "", errTargetDomainNotActive
 	}
