+++ service/matching/taskListManager.go
@@ -31,7 +31,6 @@ import (
 	"github.com/uber/cadence/common"
 	"github.com/uber/cadence/common/backoff"
 	"github.com/uber/cadence/common/cache"
-	"github.com/uber/cadence/common/domain"
 	"github.com/uber/cadence/common/log"
 	"github.com/uber/cadence/common/log/tag"
 	"github.com/uber/cadence/common/messaging"
@@ -240,7 +239,7 @@ func (c *taskListManagerImpl) AddTask(ctx context.Context, params addTaskParams)
 
 		isForwarded := params.forwardedFrom != ""
 
-		if _, err := domain.IsActive(domainEntry, c.engine.clusterMetadata); err != nil {
+		if _, err := domainEntry.IsActiveIn(c.engine.clusterMetadata.GetCurrentClusterName()); err != nil {
 			// standby task, only persist when task is not forwarded from child partition
 			syncMatch = false
 			if isForwarded {
@@ -357,7 +356,7 @@ func (c *taskListManagerImpl) getTask(ctx context.Context, maxDispatchPerSecond 
 	// value. Last poller wins if different pollers provide different values
 	c.matcher.UpdateRatelimit(maxDispatchPerSecond)
 
-	if _, err := domain.IsActive(domainEntry, c.engine.clusterMetadata); err != nil {
+	if _, err := domainEntry.IsActiveIn(c.engine.clusterMetadata.GetCurrentClusterName()); err != nil {
 		return c.matcher.PollForQuery(childCtx)
 	}
